Adds favorites to state and database
It will trigger a login prompt if a user selects favorite without logging. It will then add favorites or remove favorites depending on the movie existing in favorites or not. /favorites will display a view all button if logged in. The home page will display view favorites button if logged in.